sorry, multiple changes. Mainly: created a new class ("DatasetDict") which is more narrow in focus.
Multiset was getting to be unfocused... it was not only dealing with the nuts and bolts of handling dictionaries containing XArray Datasets, but also with the ordering of processing steps required for diagnostic recipe runs.  This new class helps separate the data structure details from the sequencing and higher level diagnostic interface commands.

In addition (sorry for multiple changes in one commit!!), this commit improves methods that set verbosity level and improves docstrings/method-ordering.