get test data from a subdirectory in a new pytest fixture
Tests are passing on local machine but failing during GitHub actions workflow. Working hypothesis is that in the test environment on GitHub, the pathing to the test data is different than on my local machine. Now I'm thinking [ hoping :) ] that a relative path that uses __file__ syntax will fix it.